# Make .Clear faster by ~20%