Using built-in `clear()` to zero-out contiguous array slices instead of zeroing elements one by one

```
name               old time/op    new time/op    delta
ClearContiguous-4     294µs ± 9%     237µs ±36%  -19.23%  (p=0.001 n=13+15)
ClearSplit-4          321µs ±11%     248µs ± 8%  -22.60%  (p=0.000 n=14+13)

name               old alloc/op   new alloc/op   delta
ClearContiguous-4     524kB ± 0%     524kB ± 0%   -0.00%  (p=0.000 n=11+12)
ClearSplit-4          524kB ± 0%     524kB ± 0%   -0.00%  (p=0.000 n=12+15)

name               old allocs/op  new allocs/op  delta
ClearContiguous-4      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ClearSplit-4           1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```